quick fix about embed

+++ Modules/Moderation.cs
@@ -10,11 +10,12 @@ namespace DiscordBot.Modules
     [Summary(":shield:")]
     public class Moderation : ModuleBase
     {
+
         [Command("purge")]
         [Summary("Purge message from channel")]
-        [RequireUserPermission(GuildPermission.Administrator)]
+        //[RequireUserPermission(GuildPermission.Administrator)]
         [RequireBotPermission(GuildPermission.ManageMessages)]
-        public async Task Purge([Summary("test")] int amount)
+        public async Task Purge(int amount)
         {
             if (amount <= 0)
             {
@@ -22,8 +23,14 @@ namespace DiscordBot.Modules
                 return;
             }
 
+            if (!(Context.User is SocketGuildUser userSend) || !userSend.GuildPermissions.ManageRoles)
+            {
+                await Utils.SendInvalidPerm(Context.User, Context.Channel);
+                return;
+            }
+
             var messages = (await Context.Channel.GetMessagesAsync(amount + 1).FlattenAsync()).ToList();
-            await ((SocketTextChannel) Context.Channel).DeleteMessagesAsync(messages);
+            await ((SocketTextChannel)Context.Channel).DeleteMessagesAsync(messages);
 
             var message =
                 await Context.Channel.SendMessageAsync($"{messages.Count} messages deleted successfully!");
@@ -39,63 +46,87 @@ namespace DiscordBot.Modules
         //[RequireUserPermission(GuildPermission.Administrator)]
         [RequireBotPermission(GuildPermission.KickMembers)]
         [RequireBotPermission(GuildPermission.BanMembers)]
-        public async Task Kick(IGuildUser userAccount)
+        public async Task Kick(SocketGuildUser userAccount, [Remainder] string reason = null)
         {
             await Context.Channel.DeleteMessageAsync(Context.Message);
 
-            if (!(Context.User is SocketGuildUser user)) return;
-
+            if (!(Context.User is SocketGuildUser userSend) || !userSend.GuildPermissions.ManageRoles ||
+                !Utils.CanInteractUser(userSend, userAccount))
+            {
+                await Utils.SendInvalidPerm(Context.User, Context.Channel);
+                return;
+            }
             var builder = new EmbedBuilder()
                 .WithTitle("Logged Information")
                 .AddField("User", $"{userAccount.Mention}")
-                .AddField("Moderator", $"{user.Mention}")
-                .AddField("Other Information", "Can join server again")
+                .AddField("Command issued by", $"{userSend.Mention}")
                 .WithDescription(
-                    $"You can't kick this {user} from {Context.Guild.Name}")
+                    $"You can't kick this {userAccount} from {Context.Guild.Name}")
                 .WithFooter($"{Context.User.Username}", Context.User.GetAvatarUrl())
                 .WithCurrentTimestamp()
                 .WithColor(new Color(54, 57, 62));
 
-            if (user.GuildPermissions.KickMembers)
+            if (userSend.GuildPermissions.KickMembers)
             {
-                await userAccount.KickAsync();
-                builder.WithDescription(
-                    $"This user has been kicked from {Context.Guild.Name} by {Context.User.Username}");
+                await userAccount.KickAsync(reason);
+                builder.AddField("Reason", $"{reason}")
+                    .AddField("Other Information", "Can join server again")
+                    .WithDescription(
+                    $"This user has been kicked from {Context.Guild.Name} by {Context.User.Username}!");
             }
 
             await Context.Channel.SendMessageAsync(null, false, builder.Build());
         }
 
         [Command("ban")]
-        [Description("Kick someone's ass")]
+        [Description("Ban someone's ass")]
         [Summary("Ban someone. Need admin perm and bot ban perm")]
         //[RequireUserPermission(GuildPermission.Administrator)]
         [RequireBotPermission(GuildPermission.KickMembers)]
         [RequireBotPermission(GuildPermission.BanMembers)]
-        public async Task Ban(IGuildUser userAccount)
+        public async Task Ban(SocketGuildUser userAccount, [Remainder] string reason = null)
         {
             await Context.Channel.DeleteMessageAsync(Context.Message);
 
-            if (!(Context.User is SocketGuildUser user)) return;
-
+            if (!(Context.User is SocketGuildUser userSend) || !userSend.GuildPermissions.ManageRoles ||
+                !Utils.CanInteractUser(userSend, userAccount))
+            {
+                await Utils.SendInvalidPerm(Context.User, Context.Channel);
+                return;
+            }
             var builder = new EmbedBuilder()
                 .WithTitle("Logged Information")
                 .AddField("User", $"{userAccount.Mention}")
-                .AddField("Invoked by", $"{user.Mention}")
+                .AddField("Command issued by", $"{userSend.Mention}")
                 .WithDescription(
-                    $"You can't ban this {user} from {Context.Guild.Name}")
+                    $"You can't ban this {userAccount} from {Context.Guild.Name}")
                 .WithFooter($"{Context.User.Username}", Context.User.GetAvatarUrl())
                 .WithCurrentTimestamp()
                 .WithColor(new Color(54, 57, 62));
 
-            if (user.GuildPermissions.BanMembers)
+            if (userSend.GuildPermissions.BanMembers)
             {
-                await userAccount.BanAsync();
-                builder.WithDescription(
-                    $"This user has been banned from {Context.Guild.Name} by {Context.User.Username}");
+                await userAccount.BanAsync(0, reason);
+                builder.AddField("Reason", $"{reason}")
+                    .AddField("Other Information", "Can't join server again")
+                    .WithDescription(
+                        $"This user has been banned from {Context.Guild.Name} by {Context.User.Username}!");
             }
 
             await Context.Channel.SendMessageAsync(null, false, builder.Build());
+        }
+
+        [Command("banserver")]
+        [Description("Ban user has mutual guilds with bot")]
+        [Summary("Ban someone share mutual guilds with bot. Need admin perm and bot ban perm")]
+        [RequireBotPermission(GuildPermission.BanMembers)]
+        public async Task BanAll(SocketGuildUser user, [Remainder] string reason = null)
+        {
+            foreach (var guilds in user.MutualGuilds)
+            {
+                await guilds.AddBanAsync(user, 1, reason: $"Banned by {Context.User}. Reason: {reason}");
+            }
+            await ReplyAsync($"{user} was successfully Banned from all servers");
         }
 
         [Command("verify", true)]
@@ -106,11 +137,15 @@ namespace DiscordBot.Modules
 
             await Context.Channel.DeleteMessageAsync(Context.Message).ConfigureAwait(false);
 
-            var role = ((IGuildUser) user)?.Guild.Roles.FirstOrDefault(x => x.Name.Equals("Verified"));
+            var role = ((IGuildUser)user)?.Guild.Roles.FirstOrDefault(x => x.Name.Equals("Verified"));
 
             var verifyUser = Context.User.Mention;
 
-            if (role == null) return;
+            if (role == null)
+            {
+                role = await Context.Guild.CreateRoleAsync("Verified", Utils.MemPermissions, null, false, false);
+            }
+
             if (user.Roles.Contains(role))
             {
                 var message = await Context.Channel.SendMessageAsync($"{verifyUser} has already verified");
@@ -121,7 +156,7 @@ namespace DiscordBot.Modules
             }
             else
             {
-                await ((SocketGuildUser) Context.User).AddRoleAsync(role);
+                await ((SocketGuildUser)Context.User).AddRoleAsync(role);
 
                 var message =
                     await Context.Channel.SendMessageAsync(
@@ -133,7 +168,7 @@ namespace DiscordBot.Modules
             }
         }
 
-        [Command("Create", true)]
+        [Command("Create")]
         [Summary("Create a new role")]
         [RequireBotPermission(GuildPermission.Administrator)]
         [RequireUserPermission(GuildPermission.ManageRoles)]
@@ -141,7 +176,7 @@ namespace DiscordBot.Modules
         {
             await Context.Channel.DeleteMessageAsync(Context.Message).ConfigureAwait(false);
 
-            if (((SocketCommandContext) Context).IsPrivate || role == null) return;
+            if (((SocketCommandContext)Context).IsPrivate || role == null) return;
             await Context.Guild.CreateRoleAsync(role, GuildPermissions.None, null, false, false);
         }
 
@@ -154,12 +189,11 @@ namespace DiscordBot.Modules
         {
             await Context.Channel.DeleteMessageAsync(Context.Message).ConfigureAwait(false);
 
-            if (((SocketCommandContext) Context).IsPrivate || msg == null) return;
+            if (((SocketCommandContext)Context).IsPrivate || msg == null) return;
             msg = msg.ToLower();
 
             var role = user.Guild.Roles.FirstOrDefault(x =>
                 x.Name.ToLower().Equals(msg) || x.Id.ToString().Equals(msg) || x.Name.ToLower().Contains(msg));
-
             var builder = new EmbedBuilder()
                 .WithTitle("Logged Information")
                 .AddField("User", $"{user.Mention}")
@@ -188,7 +222,7 @@ namespace DiscordBot.Modules
             await Context.Channel.SendMessageAsync(null, false, builder.Build());
         }
 
-        [Command("give", true)]
+        [Command("give")]
         [Description("Give a role to someone")]
         [Summary("Grant someone role. Need admin perm & bot manage role perm")]
         [RequireBotPermission(GuildPermission.ManageRoles)]
@@ -196,7 +230,7 @@ namespace DiscordBot.Modules
         public async Task AddRole(SocketGuildUser user, [Remainder] string msg)
         {
             await Context.Channel.DeleteMessageAsync(Context.Message).ConfigureAwait(false);
-            if (((SocketCommandContext) Context).IsPrivate || msg == null) return;
+            if (((SocketCommandContext)Context).IsPrivate || msg == null) return;
             msg = msg.ToLower();
 
             var role = user.Guild.Roles.FirstOrDefault(x =>
@@ -208,7 +242,6 @@ namespace DiscordBot.Modules
                 await Utils.SendInvalidPerm(Context.User, Context.Channel);
                 return;
             }
-
             var builder = new EmbedBuilder()
                 .WithTitle("Logged Information")
                 .AddField("User", $"{user.Mention}")
